+++ src/Player_NOKDPS.cpp
@@ -48,8 +48,8 @@ void Player_NOKDPS::getMoves(GameState & state, const MoveArray & moves, std::ve
 
                 if (move.index() >= state.numUnits(enemy))
                 {
-                    int e = enemy;
-                    int pl = _playerID;
+                    int e = (int)enemy;
+                    int pl = (int)_playerID;
                     printf("wtf\n");
                 }
 			}
